field-monitor/requirements.txt
playwright==1.48.0
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0
keyring==25.5.0
